Refacor code to update the ui
