+++ src/events/convert_temp.js
@@ -1,60 +1,36 @@
 import getTemperature from '../utils/temperature';
+import update from '../ui/update';
 
 const convertTemperature = (e) => {
-    const UIMinMax = document.querySelector('.min-max');
-    const UITemperature = document.querySelector('.temperature');
-    const UIMin = document.querySelector('.min');
-    const UIMax = document.querySelector('.max');
-    console.log(UITemperature, UIMin, UIMax);
+  const UITemperature = document.querySelector('.temperature');
+  const UIMin = document.querySelector('.min');
+  const UIMax = document.querySelector('.max');
+  console.log(UITemperature, UIMin, UIMax);
 
-    const current = UITemperature.textContent;
-    const currentMin = UIMin.textContent;
-    const currentMax = UIMax.textContent;
-    console.log(current, currentMin, currentMax);
+  const current = UITemperature.textContent;
+  const currentMin = UIMin.textContent;
+  const currentMax = UIMax.textContent;
+  console.log(current, currentMin, currentMax);
 
-    if (e.target.checked) {
-      const { temp, min, max } = getTemperature(
-        current,
-        currentMin,
-        currentMax,
-        'F'
-      );
+  if (e.target.checked) {
+    const { temp, min, max } = getTemperature(
+      current,
+      currentMin,
+      currentMax,
+      'F'
+    );
 
-      UITemperature.innerHTML = `
-        <span>${temp}</span>
-        <span>&deg;F</span>
-        `;
+    update(temp, min, max, 'F');
+  } else {
+    const { temp, min, max } = getTemperature(
+      current,
+      currentMin,
+      currentMax,
+      'C'
+    );
 
-      UIMinMax.innerHTML = `
-              <div className="min-max">
-                Min: <span class="min">${min}</span>
-                <span>&deg;C</span>
-                Max: <span class="max">${max}</span>
-                <span>&deg;F</span>
-              </div>
-        `;
-    } else {
-      const { temp, min, max } = getTemperature(
-        current,
-        currentMin,
-        currentMax,
-        'C'
-      );
-
-      UITemperature.innerHTML = `
-          <span>${temp}</span>
-        <span>&deg;C</span>
-          `;
-
-      UIMinMax.innerHTML = `
-              <div className="min-max">
-                Min: <span class="min">${min}</span>
-                <span>&deg;C</span>
-                Max: <span class="max">${max}</span>
-                <span>&deg;C</span>
-              </div>
-        `;
-    }
+    update(temp, min, max, 'C');
+  }
 };
 
 export default convertTemperature;
